[PS-65] refactor: 사진 업로드 로직 분리
