+++ src/main/java/com/tf4/photospot/user/application/UserService.java
@@ -7,7 +7,7 @@ import org.springframework.web.multipart.MultipartFile;
 import com.tf4.photospot.global.exception.ApiException;
 import com.tf4.photospot.global.exception.domain.AuthErrorCode;
 import com.tf4.photospot.global.exception.domain.UserErrorCode;
-import com.tf4.photospot.photo.application.PhotoService;
+import com.tf4.photospot.global.util.S3Uploader;
 import com.tf4.photospot.user.application.request.LoginUserInfo;
 import com.tf4.photospot.user.application.response.OauthLoginUserResponse;
 import com.tf4.photospot.user.application.response.UserProfileResponse;
@@ -23,17 +23,18 @@ import lombok.RequiredArgsConstructor;
 public class UserService {
 
 	private final UserRepository userRepository;
-	private final PhotoService photoService;
+	private final S3Uploader s3Uploader;
 
 	@Transactional
 	public UserProfileResponse updateProfile(Long userId, MultipartFile file, String request) {
-		String photoUrl = photoService.uploadOtherPhoto(file, request);
+		String photoUrl = s3Uploader.upload(file, request);
 		User loginUser = userRepository.findById(userId)
 			.orElseThrow(() -> new ApiException(AuthErrorCode.NOT_FOUND_USER));
 		loginUser.updateProfile(photoUrl);
 		return new UserProfileResponse(photoUrl);
 	}
 
+	// Todo : 로그인 관련 메서드 AuthService 옮기기
 	@Transactional
 	public OauthLoginUserResponse oauthLogin(String providerType, String account) {
 		return userRepository.findUserByProviderTypeAndAccount(providerType, account)
